fix(frontend): counterparty icon not loaded
